:fire: YEAHHHHHH ;) , M_Pre_Gen OK, :building_construction: Reforzando Proteccion Rutas y FIxeando errores

+++ backend/helpers/Imagen.php
@@ -78,11 +78,15 @@
         }
 
         public function removerImagen ($nombreArchivo) {
-            $remover = unlink($nombreArchivo);
-            if ($remover) {
-                return true;
+            if (file_exists($nombreArchivo)) {
+                $remover = unlink($nombreArchivo);
+                if ($remover) {
+                    return true;
+                } else {
+                    return false;
+                }
             } else {
-                return false;
+                return true;
             }
         }
     }
